Update README.md (GRADED)
I have added in more details to our readme file including the files we currently have in the repository (which will have to keep updating as we progress). Also I added a checked list/task list for the tasks we have performed, so that it can be used as a future reference for anyone trying to re-do/create our project. Additionally, I added a section with our github links so that anybody with further questions can contact us. 

@jchalissery14 Can you review if the the markdown file has all the things required as well as the Github-flavoured markdown functionalities are working/written fine including the emoji.

Thanks for your co-operation!